feat: Move `IncrNonceAuthComponent`
