+++ crates/miden-lib/src/testing/account_component/mod.rs
@@ -1,0 +1,2 @@
+mod incr_nonce;
+pub use incr_nonce::IncrNonceAuthComponent;
