+++ crates/miden-lib/src/testing/mod.rs
@@ -1,0 +1,1 @@
+pub mod account_component;
