+++ crates/miden-testing/src/mock_chain/auth.rs
@@ -3,15 +3,12 @@
 use alloc::vec::Vec;
 
 use miden_lib::account::auth::{AuthRpoFalcon512, AuthRpoFalcon512Acl, AuthRpoFalcon512AclConfig};
+use miden_lib::testing::account_component::IncrNonceAuthComponent;
 use miden_lib::transaction::TransactionKernel;
 use miden_objects::Word;
 use miden_objects::account::{AccountComponent, AuthSecretKey};
 use miden_objects::crypto::dsa::rpo_falcon512::SecretKey;
-use miden_objects::testing::account_component::{
-    ConditionalAuthComponent,
-    IncrNonceAuthComponent,
-    NoopAuthComponent,
-};
+use miden_objects::testing::account_component::{ConditionalAuthComponent, NoopAuthComponent};
 use miden_tx::auth::BasicAuthenticator;
 use rand::SeedableRng;
 use rand_chacha::ChaCha20Rng;
@@ -90,11 +87,7 @@ impl Auth {
 
                 (component, Some(authenticator))
             },
-            Auth::IncrNonce => {
-                let assembler = TransactionKernel::assembler();
-                let component = IncrNonceAuthComponent::new(assembler).unwrap();
-                (component.into(), None)
-            },
+            Auth::IncrNonce => (IncrNonceAuthComponent.into(), None),
 
             Auth::Noop => {
                 let assembler = TransactionKernel::assembler();
